split Timer into smaller widgets
