+++ lib/common/globals.dart
@@ -41,3 +41,4 @@ PackageInfo packageInfo = PackageInfo(
 
 // Quick actions
 const QuickActions quickActions = QuickActions();
+bool doScroll = false;
